% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_colors.R
\name{ft_colors}
\alias{ft_colors}
\title{FT Colors}
\usage{
ft_colors(...)
}
\arguments{
\item{...}{Names of the colors}
}
\description{
Get hex codes for ft brand colors.
}
\details{
Use \code{ft_colors()} to see the full list of colours. This list
  comes from
  \href{https://registry.origami.ft.com/components/o-colors}{Origami
  o-colors}.
}
\examples{
# Full list
ft_colors()

# Choose colors
ft_colors("paper")
ft_colors("oxford", "claret")

}
