% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_fonts.R
\name{ft_fonts}
\alias{ft_fonts}
\title{FT Fonts.}
\usage{
ft_fonts(fonts_filepath = "~/.RFonts/")
}
\arguments{
\item{font_path}{storing Financier and Metric font .woff files}
}
\description{
The \code{ft_colors} function allows the loading of FT or other fonts
}
