% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ft_palettes}
\alias{ft_palettes}
\alias{ft_main}
\alias{origami_primary}
\alias{origami_secondary}
\alias{origami_tertiary}
\alias{blacks}
\alias{whites}
\alias{clarets}
\alias{oxfords}
\alias{teals}
\title{FT Color Palettes}
\usage{
ft_main(n)

origami_primary(n)

origami_secondary(n)

origami_tertiary(n)

blacks(n)

whites(n)

clarets(n)

oxfords(n)

teals(n)
}
\arguments{
\item{n}{Number of colors required}
}
\description{
Functions to create color palettes from FT colors
}
\details{
Color palettes built from
  \href{https://registry.origami.ft.com/components/o-colors}{Origami} colors.
}
\examples{
# 10 colors from main palette
ft_main(10)

}
