% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-scales.R
\name{scale_color_ft}
\alias{scale_color_ft}
\title{Color scale constructor for FT colors}
\usage{
scale_color_ft(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{Character name of a palette in ft_o_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or scale_color_gradientn()}
}
\description{
Color scale constructor for FT colors
}
\examples{
df <- data.frame(
 x = runif(50),
 y = runif(50),
 z = rnorm(50),
 category = letters[1 + rbinom(50, 3, .25)]
 )

library(ggplot2)
# plot with discrete color scale
ggplot(df) +
  geom_point(aes(x, y, color = category)) +
  scale_color_ft()

# plot with continuous color scale
ggplot(df) +
  geom_point(aes(x, y, color = z)) +
  scale_color_ft(palette = "claret", discrete = FALSE)

}
\seealso{
\code{\link{scale_fill_ft}}
}
