% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-scales.R
\name{scale_colour_ft}
\alias{scale_colour_ft}
\title{Color scale constructor for FT colors}
\usage{
scale_colour_ft(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{Character name of a palette in ft_o_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or scale_color_gradientn()}
}
\description{
Color scale constructor for FT colors
}
