% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_colors.R
\name{ft_colors}
\alias{ft_colors}
\title{FT Colour Palettes.}
\usage{
ft_colors(palette = "main")
}
\arguments{
\item{palette}{A string.}
}
\value{
A vector of colour hex codes
}
\description{
The \code{ft_colors} function returns a different ft color palettes to use in your graphs.
Palette list is "main", "greys", "reds", "blues", "paired"
}
